requirements.txt
python-telegram-bot
httpx
tenacity
asyncio
